Data Analysis AI/backend/requirements.txt
fastapi
uvicorn
pandas
polars
scikit-learn
plotly
openai
python-multipart
astunparse
pydantic
python-dotenv
pyarrow
google-genai
